@@ -18,7 +18,7 @@ package com.stfalcon.chatkit.messages;
 
 import android.content.Context;
 import android.graphics.drawable.Drawable;
-import android.support.v4.widget.Space;
+import android.os.Build;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.AttributeSet;
@@ -27,7 +27,10 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.RelativeLayout;
+import android.widget.Space;
 import android.widget.TextView;
+
+import androidx.core.view.ViewCompat;
 
 import com.stfalcon.chatkit.R;
 
@@ -36,8 +39,9 @@ import java.lang.reflect.Field;
 /**
  * Component for input outcoming messages
  */
+@SuppressWarnings({"WeakerAccess", "unused"})
 public class MessageInput extends RelativeLayout
-        implements View.OnClickListener, TextWatcher {
+        implements View.OnClickListener, TextWatcher, View.OnFocusChangeListener {
 
     protected EditText messageInput;
     protected ImageButton messageSendButton;
@@ -47,6 +51,19 @@ public class MessageInput extends RelativeLayout
     private CharSequence input;
     private InputListener inputListener;
     private AttachmentsListener attachmentsListener;
+    private boolean isTyping;
+    private TypingListener typingListener;
+    private int delayTypingStatusMillis;
+    private Runnable typingTimerRunnable = new Runnable() {
+        @Override
+        public void run() {
+            if (isTyping) {
+                isTyping = false;
+                if (typingListener != null) typingListener.onStopTyping();
+            }
+        }
+    };
+    private boolean lastFocus;
 
     public MessageInput(Context context) {
         super(context);
@@ -107,6 +124,8 @@ public class MessageInput extends RelativeLayout
             if (isSubmitted) {
                 messageInput.setText("");
             }
+            removeCallbacks(typingTimerRunnable);
+            post(typingTimerRunnable);
         } else if (id == R.id.attachmentButton) {
             onAddAttachments();
         }
@@ -120,6 +139,14 @@ public class MessageInput extends RelativeLayout
     public void onTextChanged(CharSequence s, int start, int count, int after) {
         input = s;
         messageSendButton.setEnabled(input.length() > 0);
+        if (s.length() > 0) {
+            if (!isTyping) {
+                isTyping = true;
+                if (typingListener != null) typingListener.onStartTyping();
+            }
+            removeCallbacks(typingTimerRunnable);
+            postDelayed(typingTimerRunnable, delayTypingStatusMillis);
+        }
     }
 
     /**
@@ -128,7 +155,7 @@ public class MessageInput extends RelativeLayout
      */
     @Override
     public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-
+        //do nothing
     }
 
     /**
@@ -136,7 +163,15 @@ public class MessageInput extends RelativeLayout
      */
     @Override
     public void afterTextChanged(Editable editable) {
+        //do nothing
+    }
 
+    @Override
+    public void onFocusChange(View v, boolean hasFocus) {
+        if (lastFocus && !hasFocus && typingListener != null) {
+            typingListener.onStopTyping();
+        }
+        lastFocus = hasFocus;
     }
 
     private boolean onSubmit() {
@@ -157,22 +192,22 @@ public class MessageInput extends RelativeLayout
         this.messageInput.setTextSize(TypedValue.COMPLEX_UNIT_PX, style.getInputTextSize());
         this.messageInput.setTextColor(style.getInputTextColor());
         this.messageInput.setHintTextColor(style.getInputHintColor());
-        this.messageInput.setBackground(style.getInputBackground());
+        ViewCompat.setBackground(this.messageInput, style.getInputBackground());
         setCursor(style.getInputCursorDrawable());
 
         this.attachmentButton.setVisibility(style.showAttachmentButton() ? VISIBLE : GONE);
-        this.attachmentButton.setBackground(style.getAttachmentButtonBackground());
         this.attachmentButton.setImageDrawable(style.getAttachmentButtonIcon());
         this.attachmentButton.getLayoutParams().width = style.getAttachmentButtonWidth();
         this.attachmentButton.getLayoutParams().height = style.getAttachmentButtonHeight();
+        ViewCompat.setBackground(this.attachmentButton, style.getAttachmentButtonBackground());
 
         this.attachmentButtonSpace.setVisibility(style.showAttachmentButton() ? VISIBLE : GONE);
         this.attachmentButtonSpace.getLayoutParams().width = style.getAttachmentButtonMargin();
 
-        this.messageSendButton.setBackground(style.getInputButtonBackground());
         this.messageSendButton.setImageDrawable(style.getInputButtonIcon());
         this.messageSendButton.getLayoutParams().width = style.getInputButtonWidth();
         this.messageSendButton.getLayoutParams().height = style.getInputButtonHeight();
+        ViewCompat.setBackground(messageSendButton, style.getInputButtonBackground());
         this.sendButtonSpace.getLayoutParams().width = style.getInputButtonMargin();
 
         if (getPaddingLeft() == 0
@@ -186,30 +221,52 @@ public class MessageInput extends RelativeLayout
                     style.getInputDefaultPaddingBottom()
             );
         }
+        this.delayTypingStatusMillis = style.getDelayTypingStatus();
     }
 
     private void init(Context context) {
         inflate(context, R.layout.view_message_input, this);
 
-        messageInput = (EditText) findViewById(R.id.messageInput);
-        messageSendButton = (ImageButton) findViewById(R.id.messageSendButton);
-        attachmentButton = (ImageButton) findViewById(R.id.attachmentButton);
-        sendButtonSpace = (Space) findViewById(R.id.sendButtonSpace);
-        attachmentButtonSpace = (Space) findViewById(R.id.attachmentButtonSpace);
+        messageInput = findViewById(R.id.messageInput);
+        messageSendButton = findViewById(R.id.messageSendButton);
+        attachmentButton = findViewById(R.id.attachmentButton);
+        sendButtonSpace = findViewById(R.id.sendButtonSpace);
+        attachmentButtonSpace = findViewById(R.id.attachmentButtonSpace);
 
         messageSendButton.setOnClickListener(this);
         attachmentButton.setOnClickListener(this);
         messageInput.addTextChangedListener(this);
         messageInput.setText("");
+        messageInput.setOnFocusChangeListener(this);
     }
 
     private void setCursor(Drawable drawable) {
+        if (drawable == null) return;
+
         try {
-            Field f = TextView.class.getDeclaredField("mCursorDrawableRes");
-            f.setAccessible(true);
-            f.set(this.messageInput, drawable);
-        } catch (Exception ignore) {
+            final Field drawableResField = TextView.class.getDeclaredField("mCursorDrawableRes");
+            drawableResField.setAccessible(true);
+
+            final Object drawableFieldOwner;
+            final Class<?> drawableFieldClass;
+            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.JELLY_BEAN) {
+                drawableFieldOwner = this.messageInput;
+                drawableFieldClass = TextView.class;
+            } else {
+                final Field editorField = TextView.class.getDeclaredField("mEditor");
+                editorField.setAccessible(true);
+                drawableFieldOwner = editorField.get(this.messageInput);
+                drawableFieldClass = drawableFieldOwner.getClass();
+            }
+            final Field drawableField = drawableFieldClass.getDeclaredField("mCursorDrawable");
+            drawableField.setAccessible(true);
+            drawableField.set(drawableFieldOwner, new Drawable[]{drawable, drawable});
+        } catch (Exception ignored) {
         }
+    }
+
+    public void setTypingListener(TypingListener typingListener) {
+        this.typingListener = typingListener;
     }
 
     /**
@@ -235,5 +292,22 @@ public class MessageInput extends RelativeLayout
          * Fires when user presses 'add' button.
          */
         void onAddAttachments();
+    }
+
+    /**
+     * Interface definition for a callback to be invoked when user typing
+     */
+    public interface TypingListener {
+
+        /**
+         * Fires when user presses start typing
+         */
+        void onStartTyping();
+
+        /**
+         * Fires when user presses stop typing
+         */
+        void onStopTyping();
+
     }
 }
