@@ -17,12 +17,13 @@
 package com.stfalcon.chatkit.messages;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SimpleItemAnimator;
 import android.util.AttributeSet;
+
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.SimpleItemAnimator;
 
 import com.stfalcon.chatkit.commons.models.IMessage;
 
@@ -56,18 +57,30 @@ public class MessagesList extends RecyclerView {
     }
 
     /**
-     * Set adapter for MessagesList
+     * Sets adapter for MessagesList
      *
      * @param adapter   Adapter. Must extend MessagesListAdapter
      * @param <MESSAGE> Message model class
      */
     public <MESSAGE extends IMessage>
     void setAdapter(MessagesListAdapter<MESSAGE> adapter) {
+        setAdapter(adapter, true);
+    }
+
+    /**
+     * Sets adapter for MessagesList
+     *
+     * @param adapter       Adapter. Must extend MessagesListAdapter
+     * @param reverseLayout weather to use reverse layout for layout manager.
+     * @param <MESSAGE>     Message model class
+     */
+    public <MESSAGE extends IMessage>
+    void setAdapter(MessagesListAdapter<MESSAGE> adapter, boolean reverseLayout) {
         SimpleItemAnimator itemAnimator = new DefaultItemAnimator();
         itemAnimator.setSupportsChangeAnimations(false);
 
         LinearLayoutManager layoutManager = new LinearLayoutManager(getContext(),
-                LinearLayoutManager.VERTICAL, true);
+                LinearLayoutManager.VERTICAL, reverseLayout);
 
         setItemAnimator(itemAnimator);
         setLayoutManager(layoutManager);
